Refactored playlist auto-import support
